industries/ecommerce/agents/order-processor/requirements.txt
ShopifyAPI>=12.0.0
requests>=2.31.0
